separate engine and editor code
