+++ editor/exports/custom-pipeline.ts
@@ -2,5 +2,6 @@ export * from '../../cocos/rendering/custom/graph';
 export * from '../../cocos/rendering/custom/types';
 export * from '../../cocos/rendering/custom/layout-graph';
 export * from '../../cocos/rendering/custom/layout-graph-utils';
+export * from '../../cocos/rendering/custom/layout-graph-editor';
 export * from '../../cocos/rendering/custom/archive';
 export * from '../../cocos/rendering/custom/binary-archive';
